Fixes for GPTGenerator. Small tweaks for read_gpt_file.
Modified defaults in the generator class to use thermal_emittance which is given in m-rad per m of spot size (which is just radians). Default is 0.9mrad as per SLAC, but can be set using
fw.generator.thermal_emittance = 0.9e-3
Will auto scale for the ASTRA thermal kinetic energy (which is independant of spot size).